# backend/openapi.yaml
openapi: 3.1.0
info:
    title: Catelog API
    version: 1.0.0
servers:
    - url: /api
security:
    - bearerAuth: []
paths:
...
    /watchlists:
        get:
            summary: Get all user watchlists
            tags: [WatchLists]
            responses:
                200:
                    description: WatchLists
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/WatchList" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        post:
            summary: Create a new watchlist
            tags: [WatchLists]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [name]
                            properties:
                                name:
                                    type: string
            responses:
                200:
                    description: WatchList created
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/WatchList" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /watchlists/{listId}:
        get:
            summary: Get watchlist by ID
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: WatchList
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/WatchList" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        patch:
            summary: Update watchlist details
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                sharedWith:
                                    type: array
                                    items:
                                        type: string
            responses:
                200:
                    description: WatchList updated
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/WatchList" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        delete:
            summary: Delete a specific watchlist or transfer ownership
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: WatchList processed successfully
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /watchlists/{listId}/items:
        get:
            summary: Get items in a watchlist
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: WatchList items
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    items:
                                        type: array
                                        items: { $ref: "#/components/schemas/WatchListItem" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        post:
            summary: Add item to watchlist
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [titleId]
                            properties:
                                titleId:
                                    type: string
                                    format: uuid
            responses:
                200:
                    description: Item added
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/WatchListItem" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /watchlists/{listId}/items/{itemId}:
        patch:
            summary: Update watchlist item
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                - name: itemId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                state:
                                    type: string
                                    enum: [planned, started, finished]
                                tags:
                                    type: array
                                    items:
                                        type: string
                                personalRating:
                                    type: number
            responses:
                200:
                    description: Item updated
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        delete:
            summary: Remove item from watchlist
            tags: [WatchLists]
            parameters:
                - name: listId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                - name: itemId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: Item removed
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
        systemKey:
            type: apiKey
            in: header
            name: X-System-Key
    responses:
...
        UnauthorizedError: # 401
            description: Access token is missing or invalid
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        ForbiddenError: # 403
            description: The authenticated user does not have permission to access this resource
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        NotFoundError: # 404
            description: The requested resource was not found
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        InternalServerError: # 500
            description: There was an unexpected error
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
    schemas:
        Error:
            type: object
            properties:
                message:
                    type: string
                errors: # For 400 Validation errors
                    type: array
                    items:
                        type: object
...
        WatchList:
            type: object
            required: [id, name, ownerId, sharedWith, createdAt]
            readOnly: [id, createdAt, updatedAt]
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                ownerId:
                    type: string
                    format: uuid
                sharedWith:
                    type: array
                    items:
                        type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        WatchListItem:
            type: object
            required: [id, listId, titleId, state, addedBy, createdAt]
            readOnly: [id, createdAt, updatedAt]
            properties:
                id:
                    type: string
                    format: uuid
                listId:
                    type: string
                    format: uuid
                titleId:
                    type: string
                    format: uuid
                state:
                    type: string
                    enum: [planned, started, finished]
                tags:
                    type: array
                    items:
                        type: string
                personalRating:
                    type: number
                addedBy:
                    type: string
                sortKey:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
...