# frontend-pwa/src/api/models/WatchList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WatchList = {
    id: string;
    name: string;
    ownerId: string;
    sharedWith: Array<string>;
    createdAt: string;
    updatedAt?: string;
};